+++ R/Step3Clusters.R
@@ -1,6 +1,6 @@
 #'@title Classify the Longitudinal Data Based on the Selected Measures.
 #'
-#'@description Classifies the trajectories by applying the k-means clustering
+#'@description Classifies the trajectories by applying the k-medoids or k-means 
 #'  algorithm to the measures selected by \code{Step2Selection}.
 #'
 #'@param trajSelection object of class \code{trajSelection} as returned by
@@ -15,9 +15,7 @@
 #'@param nclusters either \code{NULL} or the desired number of clusters. If \code{NULL}, the
 #'  number of clusters is determined using the criterion chosen in \code{criterion}. Defaults to \code{NULL}.
 #'@param criterion criterion to determine the optimal number of clusters if \code{nclusters} is \code{NULL}. Either \code{"GAP"} or \code{"Calinski-Harabasz"}. Defaults to \code{"Calinski-Harabasz"}.
-#'@param K.max maximum number of clusters to be considered if \code{nclusters} is set to \code{NULL}. Defaults to \code{15}.
-#'@param boot logical. If \code{TRUE}, and if \code{"Calinski-Harabasz"} is the chosen \code{criterion}, the optimal number of clusters will be the first mode of sampling distribution of the optimal number of clusters obtained by bootstrap. Defaults to \code{FALSE}.
-#'@param R the number of bootstrap replicate if \code{boot} is set to \code{TRUE}. Defaults to \code{100}.
+#'@param K.max maximum number of clusters to be considered if \code{nclusters} is set to \code{NULL}.
 #'@param B to be passed to the \code{B} argument of
 #'  \code{\link[cluster]{clusGap}} if \code{"GAP"} is the chosen \code{criterion}.
 #'@param x object of class \code{trajClusters}.
@@ -29,15 +27,13 @@
 #'
 #'@details If \code{"GAP"} is the chosen \code{criterion} for determining the optimal number of clusters, the method described by Tibshirani et al. is implemented by the \code{\link[cluster]{clusGap}} function.
 #'
-#'Instead, if \code{"Calinski-Harabasz"} is the chosen \code{criterion}, the Calinski-Harabasz index is computed for each possible number of clusters between 2 and \code{K.max} and the optimal number of clusters is the maximizer of the Calinski-Harabasz index. Moreover, if \code{boot} is set to \code{TRUE}, then, following the guidelines suggested by Mesidor et al., a sampling distribution of the optimal number of clusters is obtained by bootstrap and the optimal number of clusters is chosen to be the (first) mode of this sampling distribution. 
+#'Instead, if \code{"Calinski-Harabasz"} is the chosen \code{criterion}, the Calinski-Harabasz index is computed for each possible number of clusters between 2 and \code{K.max} and the optimal number of clusters is the maximizer of the Calinski-Harabasz index. 
 #'
 #'@import cluster
 #'@importFrom stats kmeans na.omit qt quantile var
 #'@importFrom graphics barplot
 #'
-#'@references Miceline Mésidor, Caroline Sirois, Marc Simard, Denis Talbot, A Bootstrap Approach for Evaluating Uncertainty in the Number of Groups Identified by Latent Class Growth Models, American Journal of Epidemiology, Volume 192, Issue 11, November 2023, Pages 1896–1903, https://doi.org/10.1093/aje/kwad148
-#'
-#'Tibshirani, R., Walther, G. and Hastie, T. (2001). Estimating the number of data clusters via the Gap statistic. Journal of the Royal Statistical Society B, 63, 411–423.
+#'@references Tibshirani, R., Walther, G. and Hastie, T. (2001). Estimating the number of data clusters via the Gap statistic. Journal of the Royal Statistical Society B, 63, 411–423.
 #'
 #'Tibshirani, R., Walther, G. and Hastie, T. (2000). Estimating the number of clusters in a dataset via the Gap statistic. Technical Report. Stanford.
 #'
@@ -72,9 +68,7 @@ Step3Clusters <-
             iter.max = 100,
             nclusters = NULL,
             criterion = "Calinski-Harabasz",
-            K.max = min(15, nrow(trajSelection$selection) - 1),
-            boot = FALSE,
-            R = 100,
+            K.max = min(ceiling(sqrt(nrow(trajSelection$selection))), 10),
             B = 500
   ) {
     if (is.null(nclusters) & !(criterion %in% c("Calinski-Harabasz", "GAP"))) { 
@@ -90,7 +84,6 @@ Step3Clusters <-
     }
     GAP <- NULL
     CH <- NULL
-    CH.boot <- NULL
     
     ID <- trajSelection$selection[, 1]
     
@@ -154,31 +147,6 @@ Step3Clusters <-
       }
       
       if (criterion == "Calinski-Harabasz") {
-        if (isTRUE(boot)) {
-          CH.boot <- c()
-          for(i in 1:R){
-            s1 = trajSelection$trajMeasures
-            id = s1$measures$ID 
-            indices <- sample(nrow(s1$measures), replace=TRUE)
-            s1$measures <- s1$measures[indices, ]
-            s1$measures$ID <- id
-            s2.boot <- Step2Selection(trajMeasures = s1, num.select = s1$input$num.select, discard = s1$input$discard, select = s1$input$select)
-            d = data.frame(apply(data.frame(s2.boot$selection[,-c(1), drop = FALSE]), 2, scale))
-            CH.aux <- c()
-            for(k in 2:K.max){
-              if(algorithm == "k-medoids"){
-                CH.aux[k] <- CalinskiHarabasz(x = d, clustering = cluster::pam(x = d, k = k, cluster.only = TRUE))
-              }
-              if(algorithm == "k-means"){
-                CH.aux[k] <- CalinskiHarabasz(x = d, clustering = stats::kmeans(x = d, centers = k, iter.max = iter.max, nstart = nstart)$cluster)
-              }
-            }
-            CH.boot[i] <- which(CH.aux == max(CH.aux, na.rm = T))
-          }
-          
-          nclusters <- FirstMode(CH.boot)
-          
-        } else{
           CH <- c()
           if (algorithm == "k-medoids") {
             for(k in 2:K.max){
@@ -193,7 +161,6 @@ Step3Clusters <-
             }
             nclusters <- which(CH == max(CH, na.rm = T))
           }
-        }
       }
     } 
     
@@ -248,7 +215,6 @@ Step3Clusters <-
           selection = trajSelection$selection,
           GAP = GAP,
           CH = CH,
-          CH.boot = CH.boot,
           nclusters = nclusters,
           partition = clust.by.id,
           partition.summary = partition.summary